Various fixes

+++ libraries/Moviment/Motor.h
@@ -1,13 +1,13 @@
 #ifndef Motor_h
 #define Motor_h
 
-#include "Arduino.h"
+#include <Arduino.h>
 
 class Motor {
   public:
     Motor(byte i1, byte i2);
-    void start(byte speed, bool inverse);
-    void setSpeed(byte speed);
+    void start(uint16_t speed, bool inverse);
+    void setSpeed(uint16_t speed);
     void stop();
     void begin();
   private:
